refactored prodivers for storybook. TODO fix font error for authModal story
