+++ components/modals/AuthModal.tsx
@@ -1,16 +1,16 @@
 "use client";
 
-import { useRouter } from "next/navigation";
 import Modal from "@/components/modals/Modal";
+import useAuthModal from "@/hooks/modals/useAuthModal";
 import {
 	useSessionContext,
 	useSupabaseClient,
 } from "@supabase/auth-helpers-react";
 import { Auth } from "@supabase/auth-ui-react";
 import { ThemeSupa } from "@supabase/auth-ui-shared";
-import useAuthModal from "@/hooks/modals/useAuthModal";
-import { useEffect } from "react";
 import { Figtree } from "next/font/google";
+import { useRouter } from "next/navigation";
+import { useEffect } from "react";
 
 const font = Figtree({ subsets: ["latin"] });
 
@@ -24,8 +24,10 @@ const font = Figtree({ subsets: ["latin"] });
  * 
  * On login mode, it shows a link to login using magic link email.
  * On login mode, it shows a link to reset password.
+ * 
+ * @requires CSS variables needs to be set with a variable for --main-darken, -main, --darken
  */
-const AuthModal = () => {
+export default function AuthModal() {
 	const supabaseClient = useSupabaseClient();
 	const { session } = useSessionContext();
 
@@ -102,5 +104,3 @@ const AuthModal = () => {
 		</Modal>
 	);
 };
-
-export default AuthModal;
