+++ stories/pages/account.stories.tsx
@@ -1,9 +1,7 @@
 
 
 import AccountPresenter from '@/app/account/components/AccountPresenter';
-import UserProvider from '@/providers/UserProvider';
 import type { Meta, StoryObj } from '@storybook/react';
-import { AppRouterContext } from 'next/dist/shared/lib/app-router-context.shared-runtime';
 
 const mockUser = {
   id: "mock user",
@@ -17,31 +15,12 @@ const mockUser = {
   aud: "mock aud"
 }
 
-/** This is the mock for app router context */
-const createMockRouterContext = () => ({
-  push: () => Promise.resolve(true),
-  replace: () => Promise.resolve(true),
-  refresh: () => Promise.resolve(),
-  back: () => Promise.resolve(),
-  forward: () => Promise.resolve(),
-  prefetch: () => Promise.resolve(),
-});
-
 const meta = {
   title: 'app/account',
   component: AccountPresenter,
   args: {
     user: null
   },
-  decorators: [
-    (Story) => (
-      <AppRouterContext.Provider value={createMockRouterContext()}>
-        <UserProvider>
-          <Story />
-        </UserProvider>
-      </AppRouterContext.Provider>
-    )
-  ]
 } satisfies Meta<typeof AccountPresenter>;
 
 export default meta;
