+++ stories/pages/home.stories.tsx
@@ -1,8 +1,6 @@
 
 import { HomePresenter } from '@/app/(site)/components/HomePresenter';
-import UserProvider from '@/providers/UserProvider';
 import type { Meta, StoryObj } from '@storybook/react';
-import { AppRouterContext } from 'next/dist/shared/lib/app-router-context.shared-runtime';
 
 const mockUser = {
   id: "mock user",
@@ -16,32 +14,13 @@ const mockUser = {
   aud: "mock aud"
 }
 
-/** This is the mock for app router context */
-const createMockRouterContext = () => ({
-  push: () => Promise.resolve(true),
-  replace: () => Promise.resolve(true),
-  refresh: () => Promise.resolve(),
-  back: () => Promise.resolve(),
-  forward: () => Promise.resolve(),
-  prefetch: () => Promise.resolve(),
-});
-
 const meta = {
   title: 'app/home',
   component: HomePresenter,
   args: {
     songs: [],
     user: null
-  },
-  decorators: [
-    (Story) => (
-      <AppRouterContext.Provider value={createMockRouterContext()}>
-        <UserProvider>
-          <Story />
-        </UserProvider>
-      </AppRouterContext.Provider>
-    )
-  ]
+  }  
 } satisfies Meta<typeof HomePresenter>;
 
 export default meta;
