+++ stories/pages/liked.stories.tsx
@@ -1,9 +1,7 @@
 
 
 import LikedPresenter from '@/app/liked/components/LikedPresenter';
-import UserProvider from '@/providers/UserProvider';
 import type { Meta, StoryObj } from '@storybook/react';
-import { AppRouterContext } from 'next/dist/shared/lib/app-router-context.shared-runtime';
 
 const mockUser = {
   id: "mock user",
@@ -17,16 +15,6 @@ const mockUser = {
   aud: "mock aud"
 }
 
-/** This is the mock for app router context */
-const createMockRouterContext = () => ({
-  push: () => Promise.resolve(true),
-  replace: () => Promise.resolve(true),
-  refresh: () => Promise.resolve(),
-  back: () => Promise.resolve(),
-  forward: () => Promise.resolve(),
-  prefetch: () => Promise.resolve(),
-});
-
 const meta = {
   title: 'app/liked',
   component: LikedPresenter,
@@ -34,15 +22,6 @@ const meta = {
     songs: [],
     user: null
   },
-  decorators: [
-    (Story) => (
-      <AppRouterContext.Provider value={createMockRouterContext()}>
-        <UserProvider>
-          <Story />
-        </UserProvider>
-      </AppRouterContext.Provider>
-    )
-  ]
 } satisfies Meta<typeof LikedPresenter>;
 
 export default meta;
